+++ .eslintrc.js
@@ -3,17 +3,28 @@ module.exports = {
     browser: true,
     es2020: true,
   },
-  extends: ["airbnb-base", "prettier"],
+  extends: [
+    "airbnb-base",
+    "plugin:react/recommended",
+    "prettier",
+    "prettier/react",
+  ],
+  plugins: ["react"],
   parserOptions: {
     ecmaVersion: 11,
     sourceType: "module",
+    jsx: true,
   },
   settings: {
+    react: {
+      version: "detect",
+    },
     "import/resolver": {
       webpack: {
         config: "webpack.common.js",
       },
     },
+    "import/ignore": [/\.(scss|less|css)$/],
   },
   rules: {
     camelcase: "off",