+++ app/js/pages/raffles/show.js
@@ -1,1 +1,2 @@
 import "balloon-css";
+import "@assets/css/raffles/show.css";