+++ app/js/theme.js
@@ -1,0 +1,3 @@
+export const colors = {
+  reddit: "#ff7644",
+};